tibbleprint methods

The arguments are the same as in tibble:::print.tbl except for row.names, borrowed from base::print.data.frame, and base which triggers the original behavior. Argument passed to ... whose names match base::print.data.frame's other argument names ("digits", "quote", "right", or "max"), also trigger the original behavior.

# S3 method for data.frame
print(
  x,
  ...,
  n = NULL,
  width = NULL,
  n_extra = NULL,
  row.names = TRUE,
  base = getOption("tibbleprint.base")
)

# S3 method for data.frame
format(x, ..., n = NULL, width = NULL, n_extra = NULL)

Arguments

x
Object to format or print.

...
arguments to be passed to to base::data.frame()

n
Number of rows to show. If NULL, the default, will print all rows if less than option tibble.print_max. Otherwise, will print tibble.print_min rows.

width
Width of text output to generate. This defaults to NULL, which means use getOption("tibble.width") or (if also NULL) getOption("width"); the latter displays only the columns that fit on one screen. You can also set options(tibble.width = Inf) to override this default and always print all columns.

n_extra
Number of extra columns to print abbreviated information for, if the width is too small for the entire tibble. If NULL, the default, will print information about at most tibble.max_extra_cols extra columns.

row.names
whether to display row names

base
whether to trigger original base behavior